+++ events/node_discoverd.go
@@ -3,7 +3,7 @@ package events
 import "github.com/stampzilla/gozwave/functions"
 
 type NodeDiscoverd struct {
-	Address byte
+	Address int
 
 	functions.FuncGetNodeProtocolInfo
 }
